Anne mari2 (#8)
* obstacle

Added one obstacle

* obstacletoslider

Added my obstacle to  Slider

* Update Obstacle_oliver_3.prefab

Added script move one up so that the obstacle would be higher than ground not on top of it.

* Update MainMenu.unity

Changed the size of Help text to bigger. Changed the text little bit.

* UFO

Added an UFO. If the player collides with it then he would be abductec by aliens!